Make test classes correspond to the corresponding source class.
This is a best practice and allows one to jump between sources and tests in IDE's.
