+++ webapp/tests/Controller/Team/MiscControllerTest.php
@@ -1,11 +1,16 @@
 <?php declare(strict_types=1);
 
-namespace App\Tests;
+namespace App\Tests\Controller\Team;
 
-class TeamTest extends BaseTest
+use App\Tests\BaseTest;
+
+class MiscControllerTest extends BaseTest
 {
     protected static $roles = ['team'];
 
+    /**
+     * Test that if no user is logged in the user gets redirected to the login page
+     */
     public function testTeamRedirectToLogin()
     {
         $this->client->request('GET', '/team');
@@ -16,13 +21,22 @@ class TeamTest extends BaseTest
         $this->assertEquals('http://localhost/login', $response->getTargetUrl(), $message);
     }
 
+    /**
+     * Test the login process for teams
+     */
     public function testLogin()
     {
-        # test incorrect and correct password
+        // Make sure the suer has the correct permissions
+        $this->setupUser();
+
+        // test incorrect and correct password
         $this->loginHelper('dummy', 'foo', 'http://localhost/login', 200);
-        $this->loginHelper('dummy', 'dummy', 'http://localhost/jury', 200);
+        $this->loginHelper('dummy', 'dummy', 'http://localhost/team', 200);
     }
 
+    /**
+     * Test that the team overview page contains the correct data
+     */
     public function testTeamOverviewPage()
     {
         $this->logIn();
@@ -40,6 +54,9 @@ class TeamTest extends BaseTest
         $this->assertEquals('Clarification Requests', $h3s[2]);
     }
 
+    /**
+     * Test that it is possible to create a clorification as a team
+     */
     public function testClarificationRequest()
     {
         $this->logIn();
@@ -53,8 +70,8 @@ class TeamTest extends BaseTest
         $message = var_export($link, true);
         $this->assertEquals('http://localhost/team/clarifications/add', $link->getUri(), $message);
 
-        # Note that we would like to click the link here but we cannot do
-        # that since we have too much global state, e.g. define IS_JURY
-        # constants.
+        // Note that we would like to click the link here but we cannot do
+        // that since we have too much global state, e.g. define IS_JURY
+        // constants.
     }
 }
