+++ webapp/tests/Integration/ScoreboardIntegrationTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace App\Tests;
+namespace App\Tests\Integration;
 
 use App\Entity\Contest;
 use App\Entity\ContestProblem;
